Safari-611.1.10/LayoutTests/http/tests/security/blocked-on-redirect.html

<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.dumpResourceLoadCallbacks();
}
</script>

This test produces output in the console.  Only the initial URL should be logged.

<iframe src = "http://127.0.0.1:8000/resources/redirect.php?code=307&url=http://localhost:25/security/resources/post-done.html"></iframe>
